[block] Change scheduler to CFQ for ATA/SATA
Message-id: <1334266942-23271-3-git-send-email-vgoyal@redhat.com>
Patchwork-id: 4751
O-Subject: [kernel team] [RHEL7 PATCH 2/2] block: Change scheduler to CFQ for ATA/SATA
Bugzilla: 811016
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

BZ: 811016

In previous patch we changed default IO scheduler to deadline. For SATA
disks, CFQ should still make sense. So change it and override the default
deadline.

This is not upstream. Upstream is more geared towards desktop systems and
want to stick with CFQ and expect servers to do their own tuning.

We want to provide a better out of the box experience for majority of
our customers and hope is that this change will help. Performance team
has been asking for this change for a very long time.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>