[mm] add memory tracking hooks
Message-id: <20120921181710.11235.15242.sendpatchset@file.bos.redhat.com>
Patchwork-id: 52461
O-Subject: [RHEL7 Patch RESEND] Memory tracking for Stratus
Bugzilla: 725860
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=725860

This patch enables Stratus' fault-tolerant server product.  Our product
consists of a pair of systems connected by a proprietary bus that
operate in lockstep.  The system boots up on one system.  To bring the
second system online (either at boot time or after a failure) the state
of the first system (including memory) is copied over to the second.
In order to minimize downtime we do this copy while the system is running.
This requires that we keep track of any pages that change while a copy
pass is taking place so we can go back and copy them again.  This patch
tracks both pages that are dirtied since our last pass and pages whose
state is changed by the VM subsystem.

This patch is slightly different than the versions we have used in the
past in that it uses jump label/static key functionality to minimize its
impact on normal VM operations.  Jump labels are described in
Documentation/static-keys.txt.  Essentially it's a way to turn a test
for an unlikely code path into a NOP if the test is false.  The jump
label that we're using for this code is initialized to FALSE.  When our
memory harvester module goes to work it sets it to TRUE, and page changes
are tracked.  When it's done the label is set back to FALSE and our
tracking calls become NOPs again.  Some quick hackbench tests show a very
slight overall performance improvement: running "hackbench 100 process"
took an average time over 100 runs of 13.186 without jumplabels, 13.093 with.

A version of the memory-tracking patch has been included in RHEL since
at least RHEL4.  We have made several attempts to push this patch upstream,
but have been turned down due to the lack of general applicability of this
functionality.  Stratus even released an open-source project that does
bare-metal system migration over a network connection (lksm.sourceforge.net),
but again this was deemed to be of only niche interest.

For historical purposes, here are pointers to some of the threads discussing
memory tracking on lkml and linux-mm:

My predecessor, Kimball Murray, posting in 2006:
http://lkml.indiana.edu/hypermail/linux/kernel/0609.0/1048.html

My first attempt to post it in 2009:
http://marc.info/?l=linux-mm&m=124908754213294&w=2
http://marc.info/?l=linux-mm&m=124908779313496&w=2
http://marc.info/?l=linux-mm&m=124908854314064&w=2

My second attempt to post it in 2011 (along with a pointer to our practical use case):
http://marc.info/?l=linux-mm&m=130774795028188&w=2

During an initial review of this patchset, some Red Hat engineers suggested
investigating the possibility of using the virt migration code.  I am currently
looking at the code, however, it is very unlikely that such a kernel
modification would be accepted upstream by the RHEL7 patch deadline.  The work
would be a long-term project that will require significant kernel changes as
well as a re-engineering of Stratus' product.  The idea does have merit and I
expect to investigate it further.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>