[redhat] x86_64: Kconfig: increase NR_CPUS to 5120
Message-id: <20121024181201.GA2303@redacted.bos.redhat.com>
Patchwork-id: 4954
O-Subject: [kernel team] [RHEL7 PATCH v2] x86_64: Kconfig: increase NR_CPUS to 5120 for el7
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

As discussed previously, it is thought the maximum number of supported
CPUs during the lifetime of RHEL7 will need to be higher than 4096 for a
single system. During one of the meetings on this subject, Prarit
suggested 5120 was a good place to move the bar to.

Note, we should probably revisit CONFIG_NODES_SHIFT on x86_64 as well.

Sadly we can't just patch Kconfig to bump it based on MAXSMP, and remove
the CONFIG setting, since otherwise the Fedora choice of 128 will override
our change.

[v2 -- fix default values of NR_CPUS to match, since MAXSMP makes
 NR_CPUS unselectable.]

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>