[redhat] kernel.spec: add sign-modules script and use it to sign kernel modules
Message-id: <20130125143710.GN6538@redacted.bos.redhat.com>
Patchwork-id: 56587
O-Subject: [RHEL7 PATCH] add sign-modules script and use it to sign kernel modules
Bugzilla: 893083
RH-Acked-by: Don Zickus <dzickus@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=893083
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=5319616

The modules_sign target only signs modules for which we have a
corresponding object file in addition to the kernel module. This is...
suboptimal, because we use the same build directory for every flavour we
build, and clean it when we start the next build. As a result, the only
object files left in the build directory are from the last .config
built, and if there's large divergences between flavours, a pile of
modules will be unsigned as a result.

To rectify this, add the script Fedora uses to sign their "extras"
modules separately, and use that instead of the modules_sign target.

This fixes the case where building an s390x kernel would only sign the
modules used by the kdump flavour.

[Ideally this could be further simplified if we'd instead use the same
public/private key for all builds of a kernel rpm, instead of generating
a new key per flavour.]

[[As an aside, the reason for only signing modules if there's a .o is to
  avoid signing external modules, since we're signing modules from
  INSTALL_MOD_PATH instead of before modules_install target runs.]]

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>