[redhat] configs, x86, disable CONFIG_CLOCKSOURCE_WATCHDOG
Message-id: <1361807017-4743-4-git-send-email-prarit@redhat.com>
Patchwork-id: 57072
O-Subject: [RHEL7.0 PATCH BZ 914709 3/3] configs, x86, disable CONFIG_CLOCKSOURCE_WATCHDOG
Bugzilla: 914709
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=914709

The clocksource watchdog is code that verifies the periodicity of the clock
used for system time (tsc, hpet, etc.).  It does this by comparing reads
every clock tick and determines if the ticks have exceeded a maximum
delta.  If the value does exceed the delta, the clocksource is marked
"unstable" and a message is printed to the system console.

When the watchdog fires it markes the current clocksource as "unstable",
after which there is no option to regain the "unstable" clocksource.  If
this happens on a production system the only way to recover is a reboot.

IIUC and AFAICT, the clocksource watchdog is meant to catch old systems
which may have unstable hardware clocksources.  As the maintainer of this
code in RHEL4, RHEL5, RHEL6, and now RHEL7, I have not seen a system in at
least 5 years where the watchdog actually fires because the hardware is
bad.

I've only seen this fire in cases where the system has had interrupts
disabled for a long time, or the scheduler was borked.  In both of those
cases other warnings usually indicate a problem (softlockup, hardlockup,
scheduler warnings, etc.).  I am not disputing that these warnings are
bugs *that must be resolved separately* but it is impractical to force our
customers to reboot when they get a warning because the clocksource has
been marked unstable.

We should disable the clocksource watchdog in RHEL.  It only causes
customers to have to reboot for non-fatal warnings.

Set CONFIG_CLOCKSOURCE_WATCHDOG to off in the x86 config.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>