[redhat] kernel.spec: ensure tarball is re-built if git changes
Message-id: <20130404181831.GA28509@redacted.bos.redhat.com>
Patchwork-id: 58129
O-Subject: [RHEL7 PATCH] kernel.spec: ensure tarball is re-built if git changes
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

I heard a few complaints that people were cherry-pick-ing patches and
they weren't getting re-built tarballs, resulting in some confusion as
to why patches in git weren't in the builds and such. To solve that,
use get-tar-commit-id to fish the pax extended header out of the (if it
exists) current tarball, and compare that to which git id with think
HEAD is at.

If they're the same, we can re-use the tarball, otherwise, generate a
new one.

(This is only a problem if you don't change localversion between
 testing, so some people don't hit it at all and some do, workflow
 dependent.)

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>