[x86] fix !CONFIG_HYPERVISOR_GUEST compile
Message-id: <1369755230-28532-1-git-send-email-drjones@redhat.com>
Patchwork-id: 5089
O-Subject: [kernel team] [RHEL7.0 PATCH v2] dev-compile: fix !CONFIG_HYPERVISOR_GUEST compile
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Compiling without CONFIG_HYPERVISOR_GUEST isn't something we'd do for an
official build, but developers may turn it off on purpose, or as a
result of localmodconfig. The compile problem is RHEL-specific because
x86_hyper is used unconditionally in a RHEL-only function, and thus this
patch is also rhel-only. This patch makes sure x86_hyper is defined
(as NULL) when !CONFIG_HYPERVISOR_GUEST.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>