[redhat] Kconfig: disable DEBUG_LOCK_ALLOC on ppc64-debug for now
Message-id: <20130603145219.GB17209@redacted.bos.redhat.com>
Patchwork-id: 59705
O-Subject: [RHEL7 PATCH] Kconfig: disable DEBUG_LOCK_ALLOC on ppc64-debug for now
Bugzilla: 967447

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=967447

might_sleep inside of stop_machine_run is resulting in sporadic
failures to boot ppc64 debug kernels.

Disabling the check inside of mutex_lock_nested is not ideal, but until
the code can be fixed upstream, this is blocking testing of 3.10 release
candidates.

So, do the honourable thing and disable it, limited in the Makefile
to ppc64. Make a lot of noise while we do it too, so we remember to turn
it back on for beta.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>