[xen] xenfv: fix hangs when kdumping
Message-id: <1371127700-11618-1-git-send-email-drjones@redhat.com>
Patchwork-id: 60197
O-Subject: [RHEL7.0 PATCH] xenfv: fix hangs when kdumping
Bugzilla: 845471
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=845471
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=5899052

This is a forward port of rhel6 patch 5724ae75b22ef: "xenfv: fix
hangs when kdumping". Support for kdump within xenfv guests using
pvdrivers is still not upstream. This patch allows kdump to work,
when the dump initrd contains the drivers for the emulated devices.
When support is upstream and if we wish to backport it, then this
patch will be reverted as the first patch of that series.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>