[kernel] audit: log the audit_names record type
Message-id: <1371671240-4854-3-git-send-email-jlayton@redhat.com>
Patchwork-id: 60550
O-Subject: [RHEL7.0 PATCH 2/3] audit: log the audit_names record type
Bugzilla: 908885 953186
RH-Acked-by: Richard Guy Briggs <rgb@redhat.com>
RH-Acked-by: Eric Paris <eparis@redhat.com>

...to make it clear what the intent behind each record's operation was.

In many cases you can infer this, based on the context of the syscall
and the result. In other cases it's not so obvious. For instance, in
the case where you have a file being renamed over another, you'll have
two different records with the same filename but different inode info.
By logging this information we can clearly tell which one was created
and which was deleted.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>