[fs] gfs2: Reserve journal space for quota change in do_grow
Message-id: <2087916412.102991292.1372688695731.JavaMail.root@redhat.com>
Patchwork-id: 61011
O-Subject: [RHEL7 PATCH][bz979131] GFS2: Reserve journal space for quota change in do_grow
Bugzilla: 979131
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

Earlier today I posted a RHEL6.5 patch for this bug. Here is the
RHEL7 equivalent.

If a GFS2 file system is mounted with quotas and a file is grown
in such a way that its free blocks for the allocation are represented
in a secondary bitmap, GFS2 ran out of blocks in the transaction.
That resulted in "fatal: assertion "tr->tr_num_buf <= tr->tr_blocks".
This patch reserves extra blocks for the quota change so the
transaction has enough space.

The upstream patch is here:
http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2/bmap.c?id=a01aedfe21637c965a7046271fedfdd681eba646

rhbz#979131

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>