[fs] locks: turn the blocked_list into a hashtable
Message-id: <1373384880-430-11-git-send-email-jlayton@redhat.com>
Patchwork-id: 61337
O-Subject: [RHEL7.0 PATCH 10/14] locks: turn the blocked_list into a hashtable
Bugzilla: 976876
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Break up the blocked_list into a hashtable, using the fl_owner as a key.
This speeds up searching the hash chains, which is especially significant
for deadlock detection.

Note that the initial implementation assumes that hashing on fl_owner is
sufficient. In most cases it should be, with the notable exception being
server-side lockd, which compares ownership using a tuple of the
nlm_host and the pid sent in the lock request. So, this may degrade to a
single hash bucket when you only have a single NFS client. That will be
addressed in a later patch.

The careful observer may note that this patch leaves the file_lock_list
alone. There's much less of a case for turning the file_lock_list into a
hashtable. The only user of that list is the code that generates
/proc/locks, and it always walks the entire list.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Acked-by: J. Bruce Fields <bfields@fieldses.org>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Upstream commit: 48f74186546cd5929397856eab209ebcb5692d11

Signed-off-by: Jarod Wilson <jarod@redhat.com>