[powerpc] mm: Prevent gcc to re-read the pagetables
Message-id: <20130709185707.2067.29619.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 61370
O-Subject: [PATCH RHEL7.0 BZ947166 16/23] powerpc: Prevent gcc to re-read the pagetables
Bugzilla: 947166
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=947166

Description:
------------
GCC is very likely to read the pagetables just once and cache them in
the local stack or in a register, but it is can also decide to re-read
the pagetables. The problem is that the pagetable in those places can
change from under gcc.

With THP/hugetlbfs the pmd (and pgd for hugetlbfs giga pages) can
change under gup_fast. The pages won't be freed untill we finish
gup fast because we have irq disabled and we free these pages via
rcu callback.

Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=7888b4ddb44dccd68bc20d0dc4425707dff88c72

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>