[drm] add hotspot support for cursors
Message-id: <1373503575-16860-2-git-send-email-airlied@redhat.com>
Patchwork-id: 61528
O-Subject: [PATCH 1/2] [RHEL7 PATCH] drm: add hotspot support for cursors.
Bugzilla: 983312
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Ben Skeggs <bskeggs@redhat.com>

bugzilla: 983312
upstream: 4c813d4d759c0e6b83bfd73795e9526493556dc2

This is a core drm part of the fix, the rest is in the qxl follow on.

So it looks like for virtual hw cursors on QXL we need to inform
the "hw" device what the cursor hotspot parameters are. This
makes sense if you think the host has to draw the cursor and interpret
clicks from it. However the current modesetting interface doesn't support
passing the hotspot information from userspace.

This implements a new cursor ioctl, that takes the hotspot info as well,
userspace can try calling the new interface and if it gets -ENOSYS it means
its on an older kernel and can just fallback.

Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>