[iommu] intel, amd: Remove multifunction assumption around grouping
Message-id: <20130712212710.16569.65012.stgit@bling.home>
Patchwork-id: 61800
O-Subject: [RHEL7 kernel PATCH 01/12] iommu/{vt-d, amd}: Remove multifunction assumption around grouping
Bugzilla: 984081
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 984081
Upstream commit: c14d26905d256565052dce99c35b02470162e679

If a device is multifunction and does not have ACS enabled then we
assume that the entire package lacks ACS and use function 0 as the
base of the group.  The PCIe spec however states that components are
permitted to implement ACS on some, none, or all of their applicable
functions.  It's therefore conceivable that function 0 may be fully
independent and support ACS while other functions do not.  Instead
use the lowest function of the slot that does not have ACS enabled
as the base of the group.  This may be the current device, which is
intentional.  So long as we use a consistent algorithm, all the
non-ACS functions will be grouped together and ACS functions will
get separate groups.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Joerg Roedel <joro@8bytes.org>
Signed-off-by: Jarod Wilson <jarod@redhat.com>