[vfio] Convert type1 iommu to use rbtree
Message-id: <20130712212745.16569.91198.stgit@bling.home>
Patchwork-id: 61806
O-Subject: [RHEL7 kernel PATCH 07/12] vfio: Convert type1 iommu to use rbtree
Bugzilla: 984081
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 984081
Upstream commit: cd9b22685e4ccd728550d51fbe108c473f89df4f

We need to keep track of all the DMA mappings of an iommu container so
that it can be automatically unmapped when the user releases the file
descriptor.  We currently do this using a simple list, where we merge
entries with contiguous iovas and virtual addresses.  Using a tree for
this is a bit more efficient and allows us to use common code instead
of inventing our own.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>