[pci] Allocate only as many MSI vectors as requested by driver
Message-id: <20130717022507.29907.61686.stgit@amt.stowe>
Patchwork-id: 62775
O-Subject: [RHEL7 PATCH 15/46] PCI: Allocate only as many MSI vectors as requested by driver
Bugzilla: 984759
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

Bugzilla: 984759
https://bugzilla.redhat.com/show_bug.cgi?id=984759
Upstream Status: 65f6ae66a6fb44f614a1226c398fcb38e94b3c59

commit 65f6ae66a6fb44f614a1226c398fcb38e94b3c59
Author: Alexander Gordeev <agordeev@redhat.com>
Date:   Mon May 13 11:05:48 2013 +0200

PCI: Allocate only as many MSI vectors as requested by driver

Because of the encoding of the "Multiple Message Capable" and "Multiple
Message Enable" fields, a device can only advertise that it's capable of a
power-of-two number of vectors, and the OS can only enable a power-of-two
number.

For example, a device that's limited internally to using 18 vectors would
have to advertise that it's capable of 32.  The 14 extra vectors consume
vector numbers and IRQ descriptors even though the device can't actually
use them.

This fix introduces a 'msi_desc::nvec_used' field to address this issue.
When non-zero, it is the actual number of MSIs the device will send, as
requested by the device driver.  This value should be used by architectures
to set up and tear down only as many interrupt resources as the device will
actually use.

Note, although the existing 'msi_desc::multiple' field might seem
redundant, in fact it is not.  The number of MSIs advertised need not be
the smallest power-of-two larger than the number of MSIs the device will
send.  Thus, it is not always possible to derive the former from the
latter, so we need to keep them both to handle this case.

[bhelgaas: changelog, rename to "nvec_used"]
Signed-off-by: Alexander Gordeev <agordeev@redhat.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>