[kernel] perf: Fix interrupt handler timing harness
Message-id: <1374260506-2094-15-git-send-email-jolsa@redhat.com>
Patchwork-id: 63253
O-Subject: [PATCH RHEL7.0 BZ986216 14/20] perf: Fix interrupt handler timing harness
Bugzilla: 986216
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 986216
https://bugzilla.redhat.com/show_bug.cgi?id=986216

upstream status
===============
commit e5302920da9ef23f9d19d4e9ac85704cc25bee7a
Author: Stephane Eranian <eranian@google.com>
Date:   Fri Jul 5 00:30:11 2013 +0200

description
===========
This patch fixes a serious bug in:

  14c63f17b1fd perf: Drop sample rate when sampling is too slow

There was an misunderstanding on the API of the do_div()
macro. It returns the remainder of the division and this
was not what the function expected leading to disabling the
interrupt latency watchdog.

This patch also remove a duplicate assignment in
perf_sample_event_took().

Signed-off-by: Jarod Wilson <jarod@redhat.com>