[tools] perf/evlist: Enhance perf_evlist__start_workload()
Message-id: <1374260844-2396-72-git-send-email-jolsa@redhat.com>
Patchwork-id: 63330
O-Subject: [PATCH RHEL7.0 BZ984998 71/85] perf evlist: Enhance perf_evlist__start_workload()
Bugzilla: 984998
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 984998
https://bugzilla.redhat.com/show_bug.cgi?id=984998

upstream status
===============
commit bcf3145fbeb1bd91ad2ca67b1946077530f7bfb1
Author: Namhyung Kim <namhyung.kim@lge.com>
Date:   Wed Jun 26 16:14:15 2013 +0900

description
===========
When perf tries to start a workload, it relies on a pipe which the
workload was blocked for reading.  After closing the pipe on the parent,
the workload (child) can start the actual work via exec().

However, if another process was forked after creating a workload, this
mechanism cannot work since the other process (child) also inherits the
pipe, so that closing the pipe in parent cannot unblock the workload.
Fix it by using explicit write call can then closing it.

For similar reason, the pipe fd on parent should be marked as CLOEXEC so
that it can be closed after another child exec'ed.

Signed-off-by: Jarod Wilson <jarod@redhat.com>