[tools] perf: Revert regression in configuration of Python support
Message-id: <1374260844-2396-82-git-send-email-jolsa@redhat.com>
Patchwork-id: 63341
O-Subject: [PATCH RHEL7.0 BZ984998 81/85] perf tools: Revert regression in configuration of Python support
Bugzilla: 984998
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 984998
https://bugzilla.redhat.com/show_bug.cgi?id=984998

upstream status
===============
commit a363a9da65d253fa7354ce5fd630f4f94df934cc
Author: Michael Witten <mfwitten@gmail.com>
Date:   Wed Apr 17 02:23:16 2013 +0000

description
===========
Among other things, the following:

  commit 31160d7feab786c991780d7f0ce2755a469e0e5e
  perf tools: Fix GNU make v3.80 compatibility issue

attempts to aid the user by tapping into an existing error message,
as described in the commit message:

  ... Also fix an issue where _get_attempt was called with only
  one argument. This prevented the error message from printing
  the name of the variable that can be used to fix the problem.

or more precisely:

  -$(if $($(1)),$(call _ge_attempt,$($(1)),$(1)),$(call _ge_attempt,$(2)))
  +$(if $($(1)),$(call _ge_attempt,$($(1)),$(1)),$(call _ge_attempt,$(2),$(1)))

However, The "missing" argument was in fact missing on purpose; it's
absence is a signal that the error message should be skipped, because
the failure would be due to the default value, not any user-supplied
value.  This can be seen in how `_ge_attempt' uses `gea_err' (in the
config/utilities.mak file):

  _ge_attempt = $(if $(get-executable),$(get-executable),$(_gea_warn)$(call _gea_err,$(2)))
  _gea_warn = $(warning The path '$(1)' is not executable.)
  _gea_err  = $(if $(1),$(error Please set '$(1)' appropriately))

That is, because the argument is no longer missing, the value `$(1)'
(associated with `_gea_err') always evaluates to true, thus always
triggering the error condition that is meant to be reserved for
only the case when a user explicitly supplies an invalid value.

Concretely, the result is a regression in the Makefile's configuration
of python support; rather than gracefully disable support when the
relevant executables cannot be found according to default values, the
build process halts in error as though the user explicitly supplied
the values.

This new commit simply reverts the offending one-line change.

Signed-off-by: Jarod Wilson <jarod@redhat.com>