[fs] nfs: fix open(O_RDONLY|O_TRUNC) in NFS4.0
Message-id: <1374664135-32349-1-git-send-email-jlayton@redhat.com>
Patchwork-id: 63831
O-Subject: [PATCH] [BZ#987615] nfs: fix open(O_RDONLY|O_TRUNC) in NFS4.0
Bugzilla: 987615
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Nadav Shemer <nadav@tonian.com>

This patch fixes a relatively minor bug in how NFSv4 handles the
flag combination in $SUBJECT. When passed that, the current code
returns -EBADF, which doesn't make a lot of sense since we're
not passing in a file descriptor.

The problem is that nfs4_do_setattr has some special error handling
when ATTR_OPEN is set, but since we clear it here, it never is. The
fix is to leave ATTR_OPEN set, and fix the following check to ignore
it.

This patch is not yet in mainline, but Trond has merged it into his
for-next branch and I expect it'll make 3.12.

Fixes the reproducer in this bug:

    https://bugzilla.redhat.com/show_bug.cgi?id=987615

---------------------[snip]--------------------

nfs4_proc_setattr removes ATTR_OPEN from sattr->ia_valid, but later
nfs4_do_setattr checks for it

Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>