[fs] cifs: add a "nosharesock" mount option to force new sockets to server to be created
Message-id: <1374766419-8738-2-git-send-email-sprabhu@redhat.com>
Patchwork-id: 64482
O-Subject: [RHEL 7 bz988398 01/53] cifs: add a "nosharesock" mount option to force new sockets to server to be created
Bugzilla: 988398
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Some servers set max_vcs to 1 and actually do enforce that limit. Add a
new mount option to work around this behavior that forces a mount
request to open a new socket to the server instead of reusing an
existing one.

I'd prefer to come up with a solution that doesn't require this, so
consider this a debug patch that you can use to determine whether this
is the real problem.

Cc: Jim McDonough <jmcd@samba.org>
Cc: Steve French <smfrench@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
(cherry picked from commit a0b3df5cf1fc46ad885bbc5c9f56ff0f4877beb5)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>