[fs] move sectype to the cifs_ses instead of TCP_Server_Info
Message-id: <1374766419-8738-17-git-send-email-sprabhu@redhat.com>
Patchwork-id: 64500
O-Subject: [RHEL 7 bz988398 16/53] move sectype to the cifs_ses instead of TCP_Server_Info
Bugzilla: 988398
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Now that we track what sort of NEGOTIATE response was received, stop
mandating that every session on a socket use the same type of auth.

Push that decision out into the session setup code, and make the sectype
a per-session property. This should allow us to mix multiple sectypes on
a socket as long as they are compatible with the NEGOTIATE response.

With this too, we can now eliminate the ses->secFlg field since that
info is redundant and harder to work with than a securityEnum.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Acked-by: Pavel Shilovsky <piastry@etersoft.ru>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 3f618223dc0bdcbc8d510350e78ee2195ff93768)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>