[fs] cifs: When server doesn't provide SecurityBuffer on SMB2Negotiate pick default
Message-id: <1374766419-8738-33-git-send-email-sprabhu@redhat.com>
Patchwork-id: 64508
O-Subject: [RHEL 7 bz988398 32/53] When server doesn't provide SecurityBuffer on SMB2Negotiate pick default
Bugzilla: 988398
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

According to MS-SMB2 section 2.2.4: if no blob, client picks default which
for us will be
	ses->sectype = RawNTLMSSP;
but for time being this is also our only auth choice so doesn't matter
as long as we include this fix (which does not treat the empty
SecurityBuffer as an error as the code had been doing).
We just found a server which sets blob length to zero expecting raw so
this fixes negotiation with that server.

Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 5d875cc928aa7c95c8c1e89497a9a644f32213d4)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>