[fs] cifs: Do not set DFS flag on SMB2 open
Message-id: <1374766419-8738-36-git-send-email-sprabhu@redhat.com>
Patchwork-id: 64524
O-Subject: [RHEL 7 bz988398 35/53] [CIFS] Do not set DFS flag on SMB2 open
Bugzilla: 988398
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

If we would set SMB2_FLAGS_DFS_OPERATIONS on open we also would have
to pass the path on the Open SMB prefixed by \\server\share.
Not sure when we would need to do the augmented path (if ever) and
setting this flag breaks the SMB2 open operation since it is
illegal to send an empty path name (without \\server\share prefix)
when the DFS flag is set in the SMB open header. We could
consider setting the flag on all operations other than open
but it is safer to net set it for now.

Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit f87ab88b4065a9ef00620224c4fafadc201a430c)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>