[fs] cifs: Fix a deadlock when a file is reopened
Message-id: <1374766419-8738-54-git-send-email-sprabhu@redhat.com>
Patchwork-id: 64516
O-Subject: [RHEL 7 bz988398 53/53] CIFS: Fix a deadlock when a file is reopened
Bugzilla: 988398
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

If we request reading or writing on a file that needs to be
reopened, it causes the deadlock: we are already holding rw
semaphore for reading and then we try to acquire it for writing
in cifs_relock_file. Fix this by acquiring the semaphore for
reading in cifs_relock_file due to we don't make any changes in
locks and don't need a write access.

CC: <stable@vger.kernel.org>
Signed-off-by: Pavel Shilovsky <pshilovsky@samba.org>
Acked-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 689c3db4d57a73bee6c5ad7797fce7b54d32a87c)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>